Add ButtonStyle and refactor buttons to use it
A new `ButtonStyle` has been added to `App.xaml` with properties such as `CornerRadius`, `Margin`, `FontAttributes`, `FontSize`, `HeightRequest`, `TextColor`, and `WidthRequest`.

In `DownloadsPage.xaml`, the `WidthRequest` for the `StackLayout` has been modified to use `OnIdiom` to differentiate between `Phone` and `Default` values.

In `DownloadsPage.xaml`, the `Button` for deleting a show has been updated to use the new `ButtonStyle` instead of individual properties.

In `ShowPage.xaml`, the `Button` for downloading a show and the `Button` for canceling a download have been updated to use the new `ButtonStyle` instead of individual properties.